refactor: extract medication supply card
